+++ src/include/udp-server.h
@@ -29,7 +29,7 @@ typedef struct udp_server udp_server_t;
 
 void udp_server_process_event(udp_server_t *server);
 
-udp_server_t *udp_server_new(regex_list_t *list, const char *addr, uint16_t port);
+udp_server_t *udp_server_new(regex_list_t *list, const char *addr, unsigned int port);
 
 void udp_server_close(udp_server_t *server);
